+++ example/device_emulation.dart
@@ -7,7 +7,7 @@ void main() async {
   // Emulate an iPhone 13 Pro Max
   final context = await browser.newContext(
     options: ContextOptions(
-      viewport: {'width': 428, 'height': 926},
+      viewport: ContextOptionsViewport(width: 428, height: 926),
       userAgent:
           'Mozilla/5.0 (iPhone; CPU iPhone OS 15_0 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/15.0 Mobile/15E148 Safari/604.1',
       deviceScaleFactor: 3,